requirements.txt
# Vector DB / embeddings
qdrant-client
sentence-transformers

# PDF ingestion
pymupdf
pdfminer.six

# API / web
fastapi
uvicorn
flask
python-multipart
requests

# Utils
python-dotenv
rich
typer
numpy

# LLM
google-generativeai==0.8.3

# Azure Speech
azure-cognitiveservices-speech
